% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_similarity.R
\name{cos.sim}
\alias{cos.sim}
\title{Calculate Cosine Similarity}
\usage{
\method{cos}{sim}(i, X, query)
}
\arguments{
\item{i}{Index of the column in the fold change table to compare against.}

\item{X}{The fold change data frame.}

\item{query}{The query vector for comparison.}
}
\value{
Cosine similarity value between the query and the selected dataset.
}
\description{
This function calculates the cosine similarity between two vectors.
}
