% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_skeletalvis.R
\name{download_skeletalvis}
\alias{download_skeletalvis}
\title{Download SkeltalVis Data if Not Present}
\usage{
download_skeletalvis(destfolder = "skeletalvisData", url = NULL)
}
\arguments{
\item{url}{The URL to download the Feather file from.}

\item{destfile}{The destination file path where the Feather file will be saved.}
}
\value{
The path to the Feather file.
}
\description{
This function checks if the specified skeletalvis feather file exists in the local directory.
If it does not exist, it downloads the file from the given URL.
}
\examples{
download_skeletalvis()
}
