% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_experiment.R
\name{get_experiment}
\alias{get_experiment}
\title{This function loads the differential expression data (log2 foldchanges and adjusted p-values) for a given dataset id}
\usage{
get_experiment(skeletalvis, dataset_id)
}
\arguments{
\item{skeletalvis}{The path to the skeletalvis folder.}

\item{dataset_id}{The dataset ID to extract results for.}
}
\value{
A data frame containing differential expression results for the specified dataset ID.
}
\description{
This function loads the differential expression data (log2 foldchanges and adjusted p-values) for a given dataset id
}
\examples{
skeletalvis <- load_skeletalvis(demo=TRUE)
experiment_results <- get_experiment(skeletalvis, "GSE12860_6")
}
