% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_fold_changes.R
\name{get_gene_fold_changes}
\alias{get_gene_fold_changes}
\title{Get Results for a Gene of Interest}
\usage{
get_gene_fold_changes(feather_file, gene_symbols)
}
\arguments{
\item{feather_file}{The path to the Feather file.}

\item{gene_symbols}{The gene symbol to extract results for.}
}
\value{
A data frame containing results for the specified genes.
}
\description{
This function loads a Feather file and extracts results for a specified gene.
}
\examples{
gene_results <- get_gene_fold_changes("data/mydata.feather", "BRCA1")
}
