% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query}
\alias{query}
\title{Example gene expression dataset}
\format{
## `query`
A data frame with 42,459 rows and 2 columns:
\describe{
  \item{ID}{Human Gene Symbol}
  \item{`E-MTAB-4304_1`}{log2 fold change}
}
}
\source{
<https://www.ebi.ac.uk/biostudies/arrayexpress/studies/E-MTAB-4304>
}
\usage{
query
}
\description{
An example gene expression dataset comparing damaged versus intact osteoarthritic cartilage.
}
\examples{
head(query)
}
\keyword{datasets}
