% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano Plot of Gene Expression Data}
\usage{
volcano_plot(
  data,
  id_col,
  pval_col,
  fc_col,
  logFCThreshold = log2(1.5),
  adjPValThreshold = 0.05,
  point.size = 2,
  lab.size = 4,
  top.points = 5
)
}
\arguments{
\item{data}{A data frame containing gene expression data.}

\item{id_col}{The column name for the id of the points (genes/experiments).}

\item{pval_col}{The column name for p-values.}

\item{fc_col}{The column name for fold changes.}

\item{logFCThreshold}{The foldchange threshold to define up and down regulated points}

\item{adjPValThreshold}{The pval threshold to define up and down regulated points}

\item{point.size}{Size for the points}

\item{lab.size}{Size for the labels}

\item{top.points}{Number of top up and down regulated points to label}
}
\value{
A ggplot object representing the volcano plot.
}
\description{
Volcano Plot of Gene Expression Data
}
\examples{
volcano_plot(foldChanges,"SkeletalVisID","log2FoldChange")
}
