% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_waterfall.R
\name{waterfall_plot}
\alias{waterfall_plot}
\title{Create a Waterfall Plot for log2FoldChange Data}
\usage{
waterfall_plot(data, fdr_threshold = 0.05)
}
\arguments{
\item{data}{A data frame containing `SkeletalVisID`, `log2FoldChange`, and `FDR` columns.}

\item{fdr_threshold}{A numeric value to indicate the significance threshold for `FDR`.
Entries below this threshold are considered significant.}
}
\value{
A ggplot2 object representing the waterfall plot.
}
\description{
This function generates a waterfall plot using ggplot2, where each bar represents
the `log2FoldChange` for a unique `SkeletalVisID`, ordered by value.
}
\examples{
# Example data
data <- data.frame(
  SkeletalVisID = c("E-GEOD-69110_1", "GSE155118_1", "E-GEOD-54461_8"),
  log2FoldChange = c(7.15, 3.11, -1.28),
  FDR = c(0, 0, 2.25e-195)
)
# Create waterfall plot
waterfall_plot(data, fdr_threshold = 0.05)
}
