% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_skeletalvis.R
\name{browse_skeletalvis}
\alias{browse_skeletalvis}
\title{Browse the skeletalvis database metadata}
\usage{
browse_skeletalvis(skeletalvis)
}
\arguments{
\item{skeletalvis}{The path to the SkeletalVis data folder.}
}
\value{
The datasetID of the selected gene expression profile from the interactive table.
}
\description{
Opens an interactive table showing the experiments available for analysis.
Row can be selected to return the datasetID for the comparison of interest within that experiment.
}
\examples{
if(interactive()){
skeletalvis <- load_skeletalvis(demo=TRUE)
selected_id <- browse_skeletalvis(skeletalvis)
}

}
