% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_experiment.R
\name{get_experiment}
\alias{get_experiment}
\title{Get Results for an experiment of interest}
\usage{
get_experiment(skeletalvis, dataset_id)
}
\arguments{
\item{skeletalvis}{The path to the skeletalvis folder.}

\item{dataset_id}{The dataset ID to extract results for.}
}
\value{
A data frame containing differential expression results for the specified dataset ID.
}
\description{
This function loads the skeletalvis data for a given dataset id
}
\examples{
skeletalvis <- load_skeletalvis()
experiment_results <- get_experiment(skeletalvis, "E-MTAB-4304_1")
}
