% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_skeletalvis.R
\name{load_skeletalvis}
\alias{load_skeletalvis}
\title{Find the skeletalvis data folder and download if not present}
\usage{
load_skeletalvis()
}
\value{
The path to the skeletalvis folder.
}
\description{
This function checks the default skeletalvis data location for the presence of the expected data and metadata files.
If these files do  not exist, it downloads the files from the inbuilt URLs.
}
\examples{
load_skeletalvis()
}
