% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeletalVis-package.R
\docType{package}
\name{SkeletalVis-package}
\alias{SkeletalVis}
\alias{SkeletalVis-package}
\title{SkeletalVis: Exploration and visualisation of skeletal transcriptomics data}
\description{
Allows search and visualisation of a collection of uniformly processed skeletal transcriptomic datasets. Users can find datasets where genes of intererest are differentially expressed and find datasets with a similar gene expression pattern to their own data.
}
\keyword{internal}
