% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeletalVis-package.R
\docType{package}
\name{SkeletalVis-package}
\alias{SkeletalVis}
\alias{SkeletalVis-package}
\title{SkeletalVis: Exploration and Visualisation of Skeletal Transcriptomics Data}
\description{
Allows search and visualisation of a collection of uniformly processed skeletal transcriptomic datasets. Includes methods to identify datasets where genes of interest are differentially expressed and find datasets with a similar gene expression pattern to a query dataset Soul J, Hardingham TE, Boot-Handford RP, Schwartz JM (2019) \doi{10.1093/bioinformatics/bty947}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/soulj/SkeletalVis}
  \item Report bugs at \url{https://github.com/soulj/SkeletalVis/issues}
}

}
\author{
\strong{Maintainer}: Jamie Soul \email{jamie.soul@liverpool.ac.uk} [copyright holder]

}
\keyword{internal}
