% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano Plot of Gene Expression Data}
\usage{
volcano_plot(data, pval_col, fc_col)
}
\arguments{
\item{data}{A data frame containing gene expression data.}

\item{pval_col}{The column name for p-values.}

\item{fc_col}{The column name for fold changes.}
}
\value{
A ggplot object representing the volcano plot.
}
\description{
Volcano Plot of Gene Expression Data
}
\examples{
volcano_plot(my_data, pval_col = "p_value", fc_col = "fold_change")
}
